requirements.txt
customtkinter
python-kasa
mss
numpy
watchfiles
windows-toasts
pystray
Pillow
pywin32
winshell